Add files via upload
Taking inspiration from different animals , designed a cup holder which  is specifically designed to securely hold beverages and prevent spills while 
driving. Inspired by the orangutan’s ability to maintain balance in dynamic environments and the 
hippo’s powerful grip, this tool solves the common and frustrating problem of spilled drinks during 
car travel, providing stability and security for cups and bottles. The design enables drivers to focus 
on the road without the need to hold their drink, oAering a seamless and stress-free solution to an 
everyday issue.